fix(#109): enforce utf8 request encoding
