+++ src/java/org/eurofurence/regsys/web/servlets/page/PageServlet.java
@@ -1,19 +1,16 @@
 package org.eurofurence.regsys.web.servlets.page;
 
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.eurofurence.regsys.web.pages.Page;
 import org.eurofurence.regsys.web.servlets.HttpMethod;
+import org.eurofurence.regsys.web.servlets.base.BaseServlet;
 import org.eurofurence.regsys.web.servlets.error.ErrorServlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
 
-import java.io.IOException;
-
-public class PageServlet extends HttpServlet {
+public class PageServlet extends BaseServlet {
     Logger logger = LoggerFactory.getLogger(this.getClass());
 
     protected Page buildHandler(HttpServletRequest request, HttpServletResponse response, HttpMethod method) {
@@ -27,6 +24,7 @@ public class PageServlet extends HttpServlet {
         return page;
     }
 
+    @Override
     protected void doAnyMethod(HttpServletRequest request, HttpServletResponse response, HttpMethod method) {
         String requestId = Page.createRequestId();
         try {
@@ -52,30 +50,5 @@ public class PageServlet extends HttpServlet {
         } finally {
             MDC.clear();
         }
-    }
-
-    @Override
-    public void doGet(HttpServletRequest request, HttpServletResponse response) {
-        doAnyMethod(request, response, HttpMethod.GET);
-    }
-
-    @Override
-    public void doPost(HttpServletRequest request, HttpServletResponse response) {
-        doAnyMethod(request, response, HttpMethod.POST);
-    }
-
-    @Override
-    public void doPut(HttpServletRequest request, HttpServletResponse response) {
-        doAnyMethod(request, response, HttpMethod.PUT);
-    }
-
-    @Override
-    public void doDelete(HttpServletRequest request, HttpServletResponse response) {
-        doAnyMethod(request, response, HttpMethod.DELETE);
-    }
-
-    @Override
-    public void doOptions(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        super.doOptions(request, response);
     }
 }
